+++ packages/uic-interactor/example/uic_interactor_example.dart
@@ -1,6 +1,19 @@
-import 'package:uic_interactor/uic_interactor.dart';
+/* import 'package:uic_interactor/uic_interactor.dart';
 
-void main() {
-  var awesome = Awesome();
-  print('awesome: ${awesome.isAwesome}');
+class DivideByTwoInteractor implements ParameterizedResultInteractor<int, int> {
+  const DivideByTwoInteractor();
+
+  @override
+  Future<int> execute(int input) async {
+    return input ~/ 2;
+  }
 }
+
+Future<void> main() async {
+  const divideByTwo = DivideByTwoInteractor();
+  const value = 502;
+
+  final result = await divideByTwo(value).get();
+  print('$value ~/ 2 = $result');
+}
+ */