+++ packages/uic-interactor/lib/src/modifiers/invocation_modifier.dart
@@ -1,0 +1,5 @@
+import 'package:uic_interactor/src/invocation_event.dart';
+
+abstract interface class InvocationModifier {
+  Stream<InvocationEvent> buildStream();
+}