+++ packages/uic-interactor/lib/src/parameterized_result_interactor.dart
@@ -1,0 +1,3 @@
+abstract interface class ParameterizedResultInteractor<Input, Output> {
+  Future<Output> execute(Input input);
+}