+++ packages/uic-interactor/lib/uic_interactor.dart
@@ -1,8 +1,4 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
 library;
 
-export 'src/uic_interactor_base.dart';
-
-// TODO: Export any libraries intended for clients of this package.
+export 'src/invocation_configurator.dart';
+export 'src/parameterized_result_interactor.dart';