+++ packages/uic-interactor/CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 1.0.0
 
 - Initial version.
+- Providing base interface for any interactor writting using this library.